node: extract block reassembly and submit from submitblocklight
submitblocklight rebuilt the full block (miner coinbase + the job's cached
selection) and submitted it through the chain, all inside the RPC handler. The
SV2 template provider will do the same on a SubmitSolution, so pull it out of
the RPC layer.

Move assemble_block into kth::node::mining and add submit_block_light there,
which reassembles and organizes the block, returning the organize result. The
submitblocklight handler now parses its request and calls submit_block_light;
the reassembly and submission are no longer RPC-specific. Both build as node
core, independent of KTH_WITH_RPC.

assemble_block's unit tests move with it (mining_block_submit.cpp, tagged
[mining block_submit]) and out of the RPC-gated test block, so they run in the
default build. rpc_mining.cpp keeps the render_* tests.
