+++ src/node/test/rpc_mining.cpp
@@ -4,23 +4,11 @@
 
 #include <test_helpers.hpp>
 
-#include <memory>
-#include <vector>
-
 #include <kth/domain.hpp>
 #include <kth/node/rpc/mining.hpp>
 
 using namespace kth;
 using namespace kth::node::rpc;
-
-namespace {
-
-transaction_const_ptr make_tx(uint32_t locktime) {
-    return std::make_shared<domain::message::transaction>(
-        domain::chain::transaction{1u, locktime, {}, {}});
-}
-
-} // namespace
 
 // Start Test Suite: rpc mining tests
 
@@ -47,25 +35,6 @@ TEST_CASE("render_mining_template serializes the GBT-light fields", "[rpc mining
         R"("noncerange":"00000000ffffffff",)"
         R"("mutable":["time","transactions","prevblock"],)"
         R"("job_id":"testjob"})");
-}
-
-TEST_CASE("assemble_block puts the coinbase first then the job selection", "[rpc mining]") {
-    domain::chain::transaction coinbase{1u, 0u, {}, {}};
-    std::vector<transaction_const_ptr> job{make_tx(11), make_tx(22)};
-
-    auto const block = assemble_block(domain::chain::header{}, coinbase, job);
-
-    REQUIRE(block.transactions().size() == 3u);
-    REQUIRE(block.transactions()[0].hash() == coinbase.hash());
-    REQUIRE(block.transactions()[1].hash() == job[0]->hash());
-    REQUIRE(block.transactions()[2].hash() == job[1]->hash());
-}
-
-TEST_CASE("assemble_block with an empty selection is coinbase-only", "[rpc mining]") {
-    domain::chain::transaction coinbase{1u, 7u, {}, {}};
-    auto const block = assemble_block(domain::chain::header{}, coinbase, {});
-    REQUIRE(block.transactions().size() == 1u);
-    REQUIRE(block.transactions()[0].hash() == coinbase.hash());
 }
 
 TEST_CASE("render_mining_info serializes the getmininginfo fields", "[rpc mining]") {
